Support PHP 8.x, Blesta 5.10.x and Cerb 4.10.x (#8)
This adds support for PHP 8.x (tested on PHP 8.2 and 8.3) along with
Blesta 5.10.x (tested on Blesta 5.10.1) and Cerb 4.10.x (tested on Cerb
10.4.18).

+++ cerberus_controller.php
@@ -124,8 +124,8 @@ class CerberusController extends AppController
         $hour = 3600; // seconds in an hour
 
         $days_since = floor($time/$day); // Number of days since
-        $hours_since = ($time/$hour)%24; // Number of hours since
-        $mins_since = ($time/60)%60; // Number of mins since
+        $hours_since = (int)($time/$hour)%24; // Number of hours since
+        $mins_since = (int)($time/60)%60; // Number of mins since
 
         // Set the time language
         $days_since_lang = ($days_since > 0 ? Language::_("cerberus.time_since.day", true, $days_since) : "");
